Add containerised build tooling
This commit adds container-based tooling which permits the app to
be built without any local configuration. This also permits future
work to build and publish releases automatically.

Details:

-   Add a Visual Studio Code devcontainer which installs and sets
    up all required tools automatically. This removes the need to
    have them installed locally.
-   Add a simple Maven configuration which builds and packages the
    app. This also allows tests to be dropped in.
-   Swap from Inno Setup to wix(l), which allows the installer to
    be built from the command line, even on Linux.
-   Update the README with exact build steps, including on the
    generation of native executables and the installer.
-   Add pre-commit tooling to automatically lint any changes on
    commit.
