+++ .markdownlint.yaml
@@ -1,0 +1,15 @@
+---
+MD007:
+  indent: 4
+MD013:
+  line_length: 80
+  tables: false
+MD030:
+  ul_single: 3
+  ol_single: 2
+  ul_multi: 3
+  ol_multi: 2
+MD033:
+  allowed_elements:
+    - pre
+    - br
