--- DuckDNSClient/DuckDnsInstaller/DuckDNS.xml
@@ -1,28 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<launch4jConfig>
-  <dontWrapJar>false</dontWrapJar>
-  <headerType>gui</headerType>
-  <jar>C:\Eclipse_Projects\1.NewWorkSpace\DuckDNSClient\DuckDNSClient\DuckDnsInstaller\duckdns.jar</jar>
-  <outfile>C:\Eclipse_Projects\1.NewWorkSpace\DuckDNSClient\DuckDNSClient\DuckDnsInstaller\DuckDns.exe</outfile>
-  <errTitle></errTitle>
-  <cmdLine></cmdLine>
-  <chdir></chdir>
-  <priority>normal</priority>
-  <downloadUrl>http://java.com/download</downloadUrl>
-  <supportUrl></supportUrl>
-  <stayAlive>false</stayAlive>
-  <manifest></manifest>
-  <icon>C:\Eclipse_Projects\1.NewWorkSpace\DuckDNSClient\DuckDNSClient\DuckDnsInstaller\logo.ico</icon>
-  <singleInstance>
-    <mutexName>DuckDnsUpdater</mutexName>
-    <windowTitle>DuckDnsUpdater</windowTitle>
-  </singleInstance>
-  <jre>
-    <path></path>
-    <bundledJre64Bit>false</bundledJre64Bit>
-    <minVersion>1.0.0</minVersion>
-    <maxVersion></maxVersion>
-    <jdkPreference>preferJre</jdkPreference>
-    <runtimeBits>64/32</runtimeBits>
-  </jre>
-</launch4jConfig>
