--- DuckDNSClient/DuckDnsInstaller/DuckDnsUpdaterInstaller.iss
@@ -1,41 +1,0 @@
-; Script generated by the Inno Setup Script Wizard.
-; SEE THE DOCUMENTATION FOR DETAILS ON CREATING INNO SETUP SCRIPT FILES!
-
-[Setup]
-; NOTE: The value of AppId uniquely identifies this application.
-; Do not use the same AppId value in installers for other applications.
-; (To generate a new GUID, click Tools | Generate GUID inside the IDE.)
-AppId={{F4D03CB7-3B18-44CB-AA4A-4F83FBAEBE8A}
-AppName=DuckDns Updater
-AppVersion=1.0.2
-;AppVerName=DuckDns Updater 1.0.2
-AppPublisher=ETX Software Inc.
-AppPublisherURL=http://www.etx.ca
-AppSupportURL=http://www.etx.ca
-AppUpdatesURL=http://www.etx.ca
-DefaultDirName={pf}\DuckDNSUpdater
-DefaultGroupName=DuckDns Updater
-OutputDir=C:\Eclipse_Projects\DuckDNSClient\DuckDnsInstaller
-OutputBaseFilename=DuckDnsUpdater-Installer
-SetupIconFile=C:\Eclipse_Projects\DuckDNSClient\DuckDnsInstaller\logo.ico
-Compression=lzma
-SolidCompression=yes
-
-[Languages]
-Name: "english"; MessagesFile: "compiler:Default.isl"
-
-[Tasks]
-Name: "desktopicon"; Description: "{cm:CreateDesktopIcon}"; GroupDescription: "{cm:AdditionalIcons}"; Flags: unchecked
-Name: "quicklaunchicon"; Description: "{cm:CreateQuickLaunchIcon}"; GroupDescription: "{cm:AdditionalIcons}"; Flags: unchecked; OnlyBelowVersion: 0,6.1
-
-[Files]
-Source: "C:\Eclipse_Projects\DuckDNSClient\DuckDnsInstaller\DuckDnsUpdater.exe"; DestDir: "{app}"; Flags: ignoreversion
-; NOTE: Don't use "Flags: ignoreversion" on any shared system files
-
-[Icons]
-Name: "{group}\DuckDns Updater"; Filename: "{app}\DuckDnsUpdater.exe"
-Name: "{group}\{cm:ProgramOnTheWeb,DuckDns Updater}"; Filename: "http://www.etx.ca"
-Name: "{group}\{cm:UninstallProgram,DuckDns Updater}"; Filename: "{uninstallexe}"
-Name: "{commondesktop}\DuckDns Updater"; Filename: "{app}\DuckDnsUpdater.exe"; Tasks: desktopicon
-Name: "{userappdata}\Microsoft\Internet Explorer\Quick Launch\DuckDns Updater"; Filename: "{app}\DuckDnsUpdater.exe"; Tasks: quicklaunchicon
-Name: "{commonstartup}\DuckDnsUpdater"; Filename: "{app}\DuckDnsUpdater.exe"
