--- DuckDNSClient/DuckDnsInstaller/How to Release a New version of DuckDNS Client.txt
@@ -1,12 +1,0 @@
-How to Release a New version
-
-- first create executable jar in eclipse (export / java / executable jar)
-- open launch4j (http://launch4j.sourceforge.net) and create 'DuckDns.exe' using settings contained in: \DuckDNSClient\DuckDnsInstaller\DuckDNSUpdater.xml
-- open \DuckDNSClient\DuckDnsInstaller\DuckDnsUpdaterInstaller.iss in inno setup compiler (http://www.jrsoftware.org/isdl.php#stable) and build new installation exe: DuckDnsInst.exe
-- zip and rar if required for a web release at ETX.ca
-
-- Ask for it to be uploaded to ETX web:
-  -DuckDnsIns.exe
-  -DuckDnsInst.rar
-  -DuckDnsIns.zip
-  -duckdns.jar
