--- DuckDNSClient/DuckDnsInstaller/launch4j.log
@@ -1,19 +1,0 @@
-
-
-CmdLine:	C:\Eclipse_Projects\DuckDNSClient\DuckDnsInstaller\DuckDnsUpdater.exe --l4j-debug
-WOW64:		yes
-64-bit search:	SOFTWARE\JavaSoft\Java Runtime Environment...
-Match:		SOFTWARE\JavaSoft\Java Runtime Environment\1.6
-Match:		SOFTWARE\JavaSoft\Java Runtime Environment\1.6.0_31
-Match:		SOFTWARE\JavaSoft\Java Runtime Environment\1.7
-Match:		SOFTWARE\JavaSoft\Java Runtime Environment\1.7.0
-Using 64-bit runtime.
-64-bit search:	SOFTWARE\JavaSoft\Java Development Kit...
-Ignore:		SOFTWARE\JavaSoft\Java Development Kit\1.6
-Ignore:		SOFTWARE\JavaSoft\Java Development Kit\1.6.0_31
-Using 64-bit runtime.
-Check launcher:	C:\Program Files\Java\jre7\bin\javaw.exe (OK)
-Launcher:	C:\Program Files\Java\jre7\bin\javaw.exe
-Launcher args:	-jar "C:\Eclipse_Projects\DuckDNSClient\DuckDnsInstaller\DuckDnsUpdater.exe"
-Args length:	76/32768 chars
-Exit code:	0
